modularized some routes
